package/python-cryptography: bump to version 3.4.6
Python 2 is no longer supported, a rust toolchain is also
now a direct compile time requirement for cryptography.

Drop all dependencies that are no longer needed to build
cryptography.

Propagate the new rust toolchain and python 3 requirement
to all packages that depend on python-cryptography.

Drop the C++ toolchain dependency for packages in which
that dependency was only needed for python-cryptography.

Drop python 2 dependencies for packages that depend
on python-cryptography.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>