package/gcc: fix libsanitizer in gcc 11.1, 10.3 for linux 5.13
Starting from 5.13 Cyclades async mux support is dropped from linux kernel:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=f76edd8f7ce06cdff2fe5b6b39a49644c684a161

But gcc's libsanitizer uses <linux/cyclades.h> header.
This issue is fixed by:
https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=745dae5923aba02982563481d75a21595df22ff8

Right now just gcc 9.4.0 has this commit and 8.4.0 is proposed to be deleted.
So we need to add this patch for 10.3 and 11.1 to fix build with linux kernel 5.13.

Signed-off-by: Maxim Kochetkov <fido_max@inbox.ru>