package/qt5/qt5webengine: bump to version 5.15.8
We need to switch to the kde upstream as the existing upstream no
longer updates the qt5 branch for most packages.

We need to add a separate package for the chromium submodule so that
we can apply patches for python3 support to it and replace catapult
with a python3 compatible version.

There are python3 patches available that apply to this version with
minor modifications(to be added in a future patch).

We need to set QT5WEBENGINE_SYNC_QT_HEADERS = YES as we are building
from a repository archive instead of a release archive.

We need to add some new required dependencies to qt5webengine so that
it will build.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
---
Changes v3 -> v4:
  - linker fixes

Changes v2 -> v3:
  - add kconfig entries