package/qt5: bump packages to latest kde submodule versions
This points all qt5 packages to the latest kde submodule versions
available at https://invent.kde.org/qt/qt/qt5/-/tree/kde/5.15

We need to remove some patches applied upstream already.

We need to rework qt5location to pull in the mapboxgl dependency from
a separate package as that dependency is only included in release
archives.

We need to pin qt5 packages not available in the new upstream to the
previous 5.15.2 version.

We need to set SYNC_QT_HEADERS for all qt5 packages in the new
upstream as headers are only copied in release archives which are
not provided in the kde upstream.

Set the MODULE_VERSION to 5.15.8 to match latest qt5webengine which
should be the closest release to the kde patches that don't have
lts tags.

Create a .git file so to trigger headers sync during configure.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
---
Changes v2 -> v3:
  - add kconfig entry for qt5location-mapboxgl

Changes v1 -> v2:
  - fix headers sync and update MODULE_VERSION