package/bpftool: new package
Move the target bpftool build out of linux-tools so that it is
up to date.

This also fixes build issues due to differences between kernel
versions. The latest version should be fully backwards compatible.

The host bpftool is needed for enabling features such as the systemd
bpf-framework(to be added in a future patch).

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
---
Changes v3 -> v4:
  - remove linux-tools bpftool kconfig entry
Changes v2 -> v3:
  - fix target install
Changes v1 -> v2:
  - move target bpftool out of linux-tools into stand-alone build