package/pkg-python: add pyo3(setuptools-rust/maturin) infrastructure
The pyo3 package infrastructure provides two build backends, one
of which is based on setuptools while the other is a pep517 backend.

The setuptools-rust build backend is a setuptools extension that is
capable of building python rust extensions.

The maturin build backend is a pep517 build extension that is itself
written in rust, it is itself bootstrapped using setuptools-rust but
is not itself a setuptools extension and uses pep517 infrastructure.

As both backends use pyo3 we need to set up the pyo3 environment in
addition to cargo vendoring for each.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
---
Changes v3 -> v4:
  - generalize infrastructure to support both pyo3 build backends