package/python-installer: migrate setup type to flit
This package is moving to flit and will soon be dropping distutils
compatibility support.

We need to use flit-bootstrap as opposed to the normal flit setup
type since host-python-pypa-build depends on host-python-installer.

We need to add the src directory to the PYTHONPATH so that installer
can run from the src directory when installing itself.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>