package/gcc/gcc-bpf: add BPF target support
This adds GCC BPF target support, however unlike traditional targets
the BPF target is effectively used by packages such as systemd as
a separate language rather than an entirely separate architecture.

As such we need to build what is effectively a multi-arch toolchain
for GCC BPF support, as the BPF target is effectively a bare metal
target this isn't the same as a multi-lib toolchain and generally
avoids library conflict issues as the BPF target does not have
userspace libraries.

This adds an additional gcc build stage which is similar to gcc-final
and added as a toolchain dependency when BPF target support is
enabled.

Similar to our gcc toolchain we also need a toolchain wrapper, this
is separate from our normal toolchain wrapper.

Cc: Jose E. Marchesi <jose.marchesi@oracle.com>
Cc: David Faust <david.faust@oracle.com>
Signed-off-by: James Hilliard <james.hilliard1@gmail.com>