package/pkg-cargo: configure cargo profiles
The cargo profile flags currently only select a cargo build flag which
does not accurately reflect the configured build settings, setting the
cargo build flag also only works for packages using the cargo
infrastructure directly and not with other packages using the cargo
environment such as pyo3 based python packages(which indirectly invoke
cargo build via the setuptools-rust and maturn python build backends).

To ensure that cargo uses the correct profile settings for the build
we need to set env variables for the two root profiles(dev/release)
based on the buildroot build settings.

For reference the default cargo root profile settings are:
https://doc.rust-lang.org/cargo/reference/profiles.html#default-profiles

As applications not using cargo infrastructure do not have a standard
way of configuring a specific build profile due to the selected
profile being determined by the cargo command line flags, the only
reliable way to configure profile settings is to set both root
profiles in the environment as this will ensure that they are used
regardless of which cargo build flag(such as --debug or --release) is
used.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
Tested-by: Moritz Bitsch <moritz@h6t.eu>