package/python-frozenlist: migrate to in-tree pep517 build backend
When building with a pep517 frontend we need to use the specified
build backend as opposed to the fallback setuptools build which
only works when not building with a pep517 frontend.

This package currently builds using setuptools as we do not yet
use setuptools with a pep517 build frontend. The package contains
a setuptools fallback which only can be used when using setuptools
without a pep517 frontend as the pep517 frontend will only use the
build backend specified in the package pyproject.toml which is
hatchling and not setuptools.

The custom in tree backend depends on setuptools and expandvars,
additionally it depends on cython 3 unless disabled.

As we do not currently support cython 3 lets use the pure python
build instead.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>