package/python-simplelogging: migrate to poetry core pep517 build backend
We need to migrate package/python-simplelogging to the pep517
poetry-core backend as setuptools is not supported when building
with a pep517 frontend.

This package currently builds using setuptools as we do not yet
use setuptools with a pep517 build frontend. The package contains
a setuptools fallback which only can be used when using setuptools
without a pep517 frontend as the pep517 frontend will only use the
build backend specified in the package pyproject.toml which is
poetry-core and not setuptools.

We also need to add a patch which switches from poetry to poetry
core as we do not support the full poetry package.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>