package/shadowsocks-libev: bump to version v3.3.5-53-g9afa3cacf947f910be46b69fc5a7a1fdd02fd5e6
This package has received quite a few updates but doesn't seem to be
making releases anymore.

Quite a few of these updates will be needed for compatibility with
newer mbedtls versions for example.

Lets switch to a git describe based version for now.

Drop patches which are now upstream as well.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>